platform-table: update p13 filtering dialog
