+++ apps/docs/src/app/platform/component-docs/platform-table/platform-table-examples/platform-table-p13-example.component.html
@@ -50,6 +50,7 @@
         label="Verified"
         [filterable]="true"
         dataType="boolean"
+        [groupable]="true"
     >
     </fdp-column>
 </fdp-table>
