+++ libs/platform/src/lib/components/table/components/table-p13-dialog/filtering/filtering.component.scss
@@ -1,5 +1,6 @@
-.fdp-table-p13-sorting-dialog {
-    .filter-row__input {
+.fdp-table-p13-filtering-dialog {
+    .filter-row__input,
+    .filter-row__select {
         display: block;
         margin: 0;
     }
