+++ libs/platform/src/lib/components/table/components/table-p13-dialog/filtering/filtering.component.ts
@@ -4,114 +4,11 @@ import { BehaviorSubject, Observable } from 'rxjs';
 import { DIALOG_REF, DialogRef } from '@fundamental-ngx/core';
 
 import { CollectionFilter } from '../../../interfaces';
-import { FilterStrategy, FilterableColumnDataType, getFilterStrategiesBasedOnDataType } from '../../../enums';
 import { Resettable, RESETTABLE_TOKEN } from '../../reset-button/reset-button.component';
 
-export interface FilterableColumn {
-    label: string;
-    key: string;
-    dataType: FilterableColumnDataType;
-}
+import { FilterDialogData, FilterDialogResultData, FilterableColumn, FilterRule } from './filtering.model';
 
-export interface FilterDialogData {
-    collectionFilter: CollectionFilter[];
-    columns: FilterableColumn[];
-}
-
-export interface FilterDialogResultData {
-    collectionFilter: CollectionFilter[];
-}
-
-export class FilterRule<T = any> {
-    /** Validation flg */
-    isValid = false;
-
-    /** Available strategies options */
-    strategies: ReadonlyArray<FilterStrategy> = [];
-
-    /** Data type */
-    dataType: FilterableColumnDataType;
-
-    constructor(
-        readonly columns: ReadonlyArray<FilterableColumn>,
-        /** Column key the rule belongs to */
-        public columnKey?: string,
-        /** Data type */
-        public strategy?: FilterStrategy,
-        /** Main filter value */
-        public value?: T,
-        /** Additional filter value */
-        public value2?: T
-    ) {
-        if (!this.columnKey) {
-            this.setColumnKey(columns[0]?.key);
-        }
-        if (!this.dataType) {
-            this.setDataTypeByColumnKey(this.columnKey);
-        }
-        if (this.strategies.length === 0) {
-            this.setStrategiesByColumnKey(this.columnKey);
-        }
-    }
-
-    setValid(isValid: boolean): void {
-        this.isValid = isValid;
-    }
-
-    setValue(value: T): void {
-        this.value = value;
-    }
-
-    setValue2(value: T): void {
-        this.value2 = value;
-    }
-
-    setStrategy(strategy: FilterStrategy): void {
-        this.strategy = strategy;
-    }
-
-    setStrategiesByColumnKey(columnKey: string): void {
-        const dataType = this.columns.find((column) => column.key === columnKey)?.dataType;
-        const strategies = getFilterStrategiesBasedOnDataType(dataType);
-
-        if (this.strategies === strategies) {
-            return;
-        }
-
-        this.strategies = strategies;
-
-        if (!this.strategies.includes(this.strategy)) {
-            this.setStrategy(strategies[0]);
-        }
-    }
-
-    setColumnKey(columnKey: string): void {
-        if (columnKey === this.columnKey) {
-            return;
-        }
-        this.columnKey = columnKey;
-
-        // reset values
-        this.setValue(null);
-        this.setValue2(null);
-
-        // update data type
-        this.setDataTypeByColumnKey(columnKey);
-
-        // update available strategies list
-        this.setStrategiesByColumnKey(columnKey);
-    }
-
-    setDataTypeByColumnKey(columnKey: string): void {
-        const dataType = this.columns.find((column) => column.key === columnKey).dataType;
-
-        if (dataType === this.dataType) {
-            return;
-        }
-
-        this.dataType = dataType;
-    }
-}
+export { FilterDialogData, FilterDialogResultData };
 
 @Component({
     templateUrl: './filtering.component.html',
@@ -121,31 +18,32 @@ export class FilterRule<T = any> {
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class P13FilteringComponent implements Resettable {
-    /** Table columns available for filtering */
-    readonly columns: FilterableColumn[] = [];
-
     /** @hidden */
     private _isResetAvailableSubject$: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
     /** Indicates when reset command is available */
     readonly isResetAvailable$: Observable<boolean> = this._isResetAvailableSubject$.asObservable();
 
+    /** Table columns available for filtering */
+    readonly columns: FilterableColumn[] = [];
+
     /** Initial filterBy collection */
-    initialCollectionFilter: CollectionFilter[];
+    readonly initialCollectionFilter: CollectionFilter[];
 
     /** Include Rules */
-    includeRules: FilterRule[] = [];
+    _includeRules: FilterRule[] = [];
 
     /** Exclude Rules */
-    excludeRules: FilterRule[] = [];
+    _excludeRules: FilterRule[] = [];
 
     /** Count of valid included rules */
-    validIncludeRulesCount = 0;
+    _validIncludeRulesCount = 0;
     /** Count of valid excluded rules */
-    validExcludeRulesCount = 0;
+    _validExcludeRulesCount = 0;
 
-    includePanelExpanded = true;
-
-    excludePanelExpanded = false;
+    /** Panel opened/closed flag  */
+    _includePanelExpanded = true;
+    /** Panel opened/closed flag  */
+    _excludePanelExpanded = false;
 
     constructor(@Inject(DIALOG_REF) public dialogRef: DialogRef) {
         const { columns, collectionFilter }: FilterDialogData = this.dialogRef.data;
@@ -158,8 +56,8 @@ export class P13FilteringComponent implements Resettable {
 
         this._calculateValidRulesCount();
 
-        if (this.validExcludeRulesCount && !this.validIncludeRulesCount) {
-            this.excludePanelExpanded = true;
+        if (this._validExcludeRulesCount && !this._validIncludeRulesCount) {
+            this._excludePanelExpanded = true;
         }
     }
 
@@ -222,10 +120,10 @@ export class P13FilteringComponent implements Resettable {
 
     /** @hidden */
     private _initiateRules(): void {
-        this.includeRules = this._createRules(this.initialCollectionFilter.filter(({ exclude }) => !exclude));
-        this.excludeRules = this._createRules(this.initialCollectionFilter.filter(({ exclude }) => exclude));
+        this._includeRules = this._createRules(this.initialCollectionFilter.filter(({ exclude }) => !exclude));
+        this._excludeRules = this._createRules(this.initialCollectionFilter.filter(({ exclude }) => exclude));
 
-        [this.includeRules, this.excludeRules].forEach((rules) => {
+        [this._includeRules, this._excludeRules].forEach((rules) => {
             // Rules on initial phase are considered as valid
             rules.forEach((rule) => rule.setValid(true));
             // Keep at least one item in the list
@@ -258,7 +156,7 @@ export class P13FilteringComponent implements Resettable {
 
     /** @hidden */
     private _getCollectionFiltersFromIncludeRules(): CollectionFilter[] {
-        return this._getCollectionFiltersFromRules(this.includeRules).map(
+        return this._getCollectionFiltersFromRules(this._includeRules).map(
             (collectionFilter): CollectionFilter => ({
                 ...collectionFilter,
                 exclude: false
@@ -268,7 +166,7 @@ export class P13FilteringComponent implements Resettable {
 
     /** @hidden */
     private _getCollectionFiltersFromExcludeRules(): CollectionFilter[] {
-        return this._getCollectionFiltersFromRules(this.excludeRules).map(
+        return this._getCollectionFiltersFromRules(this._excludeRules).map(
             (collectionFilter): CollectionFilter => ({
                 ...collectionFilter,
                 exclude: true
@@ -278,8 +176,8 @@ export class P13FilteringComponent implements Resettable {
 
     /** @hidden */
     private _calculateValidRulesCount = (): void => {
-        this.validIncludeRulesCount = this.includeRules.filter(this._isRuleValid).length;
-        this.validExcludeRulesCount = this.excludeRules.filter(this._isRuleValid).length;
+        this._validIncludeRulesCount = this._includeRules.filter(this._isRuleValid).length;
+        this._validExcludeRulesCount = this._excludeRules.filter(this._isRuleValid).length;
     };
 
     /** @hidden */
