+++ libs/platform/src/lib/components/table/components/table-p13-dialog/filtering/filtering.model.ts
@@ -1,0 +1,114 @@
+import { CollectionFilter } from '../../../interfaces';
+import { FilterStrategy, FilterableColumnDataType, getFilterStrategiesBasedOnDataType } from '../../../enums';
+
+export interface FilterableColumn {
+    label: string;
+    key: string;
+    dataType: FilterableColumnDataType;
+}
+
+export interface FilterDialogData {
+    collectionFilter: CollectionFilter[];
+    columns: FilterableColumn[];
+}
+
+export interface FilterDialogResultData {
+    collectionFilter: CollectionFilter[];
+}
+
+/**
+ * Filter Rule
+ *
+ * represents one rule per row
+ */
+
+export class FilterRule<T = any> {
+    /** Validation flg */
+    isValid = false;
+
+    /** Available strategies options */
+    strategies: ReadonlyArray<FilterStrategy> = [];
+
+    /** Data type */
+    dataType: FilterableColumnDataType;
+
+    constructor(
+        readonly columns: ReadonlyArray<FilterableColumn>,
+        /** Column key the rule belongs to */
+        public columnKey?: string,
+        /** Data type */
+        public strategy?: FilterStrategy,
+        /** Main filter value */
+        public value?: T,
+        /** Additional filter value */
+        public value2?: T
+    ) {
+        if (!this.columnKey) {
+            this.setColumnKey(columns[0]?.key);
+        }
+        if (!this.dataType) {
+            this.setDataTypeByColumnKey(this.columnKey);
+        }
+        if (this.strategies.length === 0) {
+            this.setStrategiesByColumnKey(this.columnKey);
+        }
+    }
+
+    setValid(isValid: boolean): void {
+        this.isValid = isValid;
+    }
+
+    setValue(value: T): void {
+        this.value = value;
+    }
+
+    setValue2(value: T): void {
+        this.value2 = value;
+    }
+
+    setStrategy(strategy: FilterStrategy): void {
+        this.strategy = strategy;
+    }
+
+    setStrategiesByColumnKey(columnKey: string): void {
+        const dataType = this.columns.find((column) => column.key === columnKey)?.dataType;
+        const strategies = getFilterStrategiesBasedOnDataType(dataType);
+
+        if (this.strategies === strategies) {
+            return;
+        }
+
+        this.strategies = strategies;
+
+        if (!this.strategies.includes(this.strategy)) {
+            this.setStrategy(strategies[0]);
+        }
+    }
+
+    setColumnKey(columnKey: string): void {
+        if (columnKey === this.columnKey) {
+            return;
+        }
+        this.columnKey = columnKey;
+
+        // reset values
+        this.setValue(null);
+        this.setValue2(null);
+
+        // update data type
+        this.setDataTypeByColumnKey(columnKey);
+
+        // update available strategies list
+        this.setStrategiesByColumnKey(columnKey);
+    }
+
+    setDataTypeByColumnKey(columnKey: string): void {
+        const dataType = this.columns.find((column) => column.key === columnKey).dataType;
+
+        if (dataType === this.dataType) {
+            return;
+        }
+
+        this.dataType = dataType;
+    }
+}
